\name{testset}
\alias{testset}
\docType{data}
\title{
De-identified test set
}
\description{
De-identified and scrambled test set to serve as the \code{rawdata} argument for \code{prepdata}. Five SNPs and 200 individuals.
}
\usage{data(testset)}
\format{
A data frame with 5 observations and 1801 variables.
}
\examples{
data(testset)
}
\keyword{testdata}
\references{Stephen W. Erickson, Joshua Callaway (2016). SNPMClust: Bivariate Gaussian Genotype Clustering and Calling for Illumina Microarrays. Journal of Statistical Software, 71(2), 1-9. doi:10.18637/jss.v071.c02}
