\name{testset}
\alias{testset}
\docType{data}
\title{
De-identified test set
}
\description{
De-identified and scrambled test set to serve as the \code{rawdata} argument for \code{prepdata}. Five SNPs and 200 individuals.
}
\usage{data(testset)}
\format{
A data frame with 5 observations and 1801 variables.
}
\examples{
data(testset)
}
\keyword{testdata}
